food-app-project/server/requirements.txt
asgiref==3.11.0
cfgv==3.5.0
distlib==0.4.0
Django==6.0.1
djangorestframework==3.16.1
filelock==3.25.2
gunicorn==25.1.0
identify==2.6.18
nodeenv==1.10.0
packaging==26.0
platformdirs==4.9.4
pre_commit==4.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-discovery==1.2.0
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
virtualenv==21.2.0
Django==6.0.1
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
psycopg==3.3.2
psycopg-binary==3.3.2
sqlparse==0.5.5
typing_extensions==4.15.0
